Move OrganizationCard to common
