+++ kifi-backend/modules/shoebox/app/com/keepit/model/OrganizationInfo.scala
@@ -73,28 +73,6 @@ object OrganizationView {
   }
 }
 
-// OrganizationCard should ONLY contain public information. No internal ids.
-case class OrganizationCard(
-  orgId: PublicId[Organization],
-  ownerId: ExternalId[User],
-  handle: OrganizationHandle,
-  name: String,
-  description: Option[String],
-  avatarPath: Option[ImagePath],
-  numMembers: Int,
-  numLibraries: Int)
-object OrganizationCard {
-  implicit val defaultWrites: Writes[OrganizationCard] = (
-    (__ \ 'id).write[PublicId[Organization]] and
-    (__ \ 'ownerId).write[ExternalId[User]] and
-    (__ \ 'handle).write[OrganizationHandle] and
-    (__ \ 'name).write[String] and
-    (__ \ 'description).writeNullable[String] and
-    (__ \ 'avatarPath).writeNullable[ImagePath] and
-    (__ \ 'numMembers).write[Int] and
-    (__ \ 'numLibraries).write[Int]
-  )(unlift(OrganizationCard.unapply))
-}
 
 object OrganizationNotificationInfoBuilder {
   def fromOrganization(org: Organization, image: Option[OrganizationAvatar])(implicit config: PublicIdConfiguration): OrganizationNotificationInfo = {
